# API/src/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Fitness API",
    "description": "A comprehensive fitness tracking and coaching API that allows coaches to manage trainees, create workout and nutrition plans, and track progress.",
    "contact": {
      "name": "API Support",
      "email": "support@fitness-api.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication"
      }
    },
    "schemas": {
...
      "SuccessMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "Food": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Name of the food item",
            "example": "Grilled Chicken Breast"
          },
          "quantity": {
            "type": "string",
            "minLength": 1,
            "description": "Quantity of the food",
            "example": "150g"
          },
          "calories": {
            "type": "number",
            "minimum": 0,
            "description": "Calories per serving",
            "example": 231
          },
          "protein": {
            "type": "number",
            "minimum": 0,
            "description": "Protein content in grams",
            "example": 43.5
          },
          "carbs": {
            "type": "number",
            "minimum": 0,
            "description": "Carbohydrate content in grams",
            "example": 0
          },
          "fat": {
            "type": "number",
            "minimum": 0,
            "description": "Fat content in grams",
            "example": 5
          },
          "fiber": {
            "type": "number",
            "minimum": 0,
            "description": "Fiber content in grams",
            "example": 0
          }
        },
        "required": [
          "name",
          "quantity",
          "calories"
        ]
      },
      "Meal": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Name of the meal",
            "example": "Breakfast"
          },
          "time": {
            "type": "string",
            "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$",
            "description": "Time when meal should be consumed",
            "example": "08:00:00"
          },
          "calories": {
            "type": "number",
            "minimum": 0,
            "description": "Total calories for the meal",
            "example": 450
          },
          "protein": {
            "type": "number",
            "minimum": 0,
            "description": "Total protein for the meal",
            "example": 25.5
          },
          "carbs": {
            "type": "number",
            "minimum": 0,
            "description": "Total carbs for the meal",
            "example": 45
          },
          "fat": {
            "type": "number",
            "minimum": 0,
            "description": "Total fat for the meal",
            "example": 12
          },
          "fiber": {
            "type": "number",
            "minimum": 0,
            "description": "Total fiber for the meal",
            "example": 8
          },
          "foods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Food"
            },
            "minItems": 1,
            "description": "List of foods in this meal"
          }
        },
        "required": [
          "name",
          "time",
          "foods"
        ]
      },
      "WeekdayPlan": {
        "type": "object",
        "properties": {
          "weekday": {
            "type": "string",
            "enum": [
              "sun",
              "mon",
              "tue",
              "wed",
              "thu",
              "fri",
              "sat"
            ],
            "description": "Day of the week",
            "example": "mon"
          },
          "totalCalories": {
            "type": "number",
            "minimum": 0,
            "description": "Total calories for the day",
            "example": 2000
          },
          "protein": {
            "type": "number",
            "minimum": 0,
            "description": "Total protein for the day",
            "example": 150
          },
          "carbs": {
            "type": "number",
            "minimum": 0,
            "description": "Total carbs for the day",
            "example": 200
          },
          "fat": {
            "type": "number",
            "minimum": 0,
            "description": "Total fat for the day",
            "example": 66
          },
          "fiber": {
            "type": "number",
            "minimum": 0,
            "description": "Total fiber for the day",
            "example": 25
          },
          "meals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Meal"
            },
            "minItems": 1,
            "description": "List of meals for this day"
          }
        },
        "required": [
          "weekday",
          "meals"
        ]
      },
      "CreateNutritionPlan": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Name of the nutrition plan",
            "example": "High Protein Diet"
          },
          "description": {
            "type": "string",
            "description": "Description of the nutrition plan",
            "example": "A diet plan focused on muscle building"
          },
          "weekdayPlans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeekdayPlan"
            },
            "minItems": 1,
            "description": "Nutrition plans for each day of the week"
          }
        },
        "required": [
          "name",
          "weekdayPlans"
        ]
      },
      "UpdateNutritionPlan": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Updated name of the nutrition plan"
          },
          "description": {
            "type": "string",
            "description": "Updated description of the nutrition plan"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the nutrition plan is active",
            "example": true
          },
          "weekdayPlans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeekdayPlan"
            },
            "description": "Updated weekday plans"
          }
        }
      },
...
      "NutritionPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Nutrition plan ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Plan name",
            "example": "High Protein Diet"
          },
          "description": {
            "type": "string",
            "description": "Plan description"
          },
          "createdBy": {
            "type": "string",
            "description": "Creator user ID"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether plan is active"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation date"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "name",
          "createdBy",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateWorkoutPlan": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Name of the workout plan",
            "example": "Push Pull Legs"
          },
          "description": {
            "type": "string",
            "description": "Description of the workout plan",
            "example": "A 6-day split focusing on push, pull, and leg movements"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "description": "Difficulty level of the workout",
            "example": "intermediate"
          },
          "estimatedCalories": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Estimated calories burned per session",
            "example": 350
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateWorkoutPlan": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Updated name of the workout plan"
          },
          "description": {
            "type": "string",
            "description": "Updated description of the workout plan"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "description": "Updated difficulty level"
          },
          "estimatedCalories": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Updated estimated calories burned"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the workout plan is active",
            "example": true
          }
        }
      },
...
      "WorkoutPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Workout plan ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Plan name",
            "example": "Push Pull Legs"
          },
          "description": {
            "type": "string",
            "description": "Plan description"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ]
          },
          "estimatedCalories": {
            "type": "number",
            "description": "Estimated calories per session"
          },
          "createdBy": {
            "type": "string",
            "description": "Creator user ID"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether plan is active"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation date"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "name",
          "createdBy",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/nutrition": {
      "get": {
        "tags": [
          "Nutrition"
        ],
        "summary": "Get all nutrition plans",
        "description": "Retrieve nutrition plans, optionally filtered by current user",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Filter by current user resources",
              "example": "true"
            },
            "required": false,
            "description": "Filter by current user resources",
            "name": "my",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of nutrition plans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NutritionPlan"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Nutrition"
        ],
        "summary": "Create nutrition plan",
        "description": "Create a new nutrition plan",
        "requestBody": {
          "description": "Nutrition plan data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNutritionPlan"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Nutrition plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionPlan"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/api/nutrition/{id}": {
      "get": {
        "tags": [
          "Nutrition"
        ],
        "summary": "Get nutrition plan by ID",
        "description": "Retrieve a specific nutrition plan",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Unique identifier",
              "example": "123"
            },
            "required": true,
            "description": "Unique identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Nutrition plan details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionPlan"
                }
              }
            }
          },
          "404": {
            "description": "Nutrition plan not found"
          }
        }
      },
      "put": {
        "tags": [
          "Nutrition"
        ],
        "summary": "Update nutrition plan",
        "description": "Update an existing nutrition plan (coach only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Unique identifier",
              "example": "123"
            },
            "required": true,
            "description": "Unique identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "description": "Updated nutrition plan data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNutritionPlan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nutrition plan updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionPlan"
                }
              }
            }
          },
          "404": {
            "description": "Nutrition plan not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Nutrition"
        ],
        "summary": "Delete nutrition plan",
        "description": "Delete a nutrition plan (coach only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Unique identifier",
              "example": "123"
            },
            "required": true,
            "description": "Unique identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Nutrition plan deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                }
              }
            }
          },
          "404": {
            "description": "Nutrition plan not found"
          }
        }
      }
    },
    "/api/workouts": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Get all workout plans",
        "description": "Retrieve workout plans, optionally filtered by current user",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Filter by current user resources",
              "example": "true"
            },
            "required": false,
            "description": "Filter by current user resources",
            "name": "my",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of workout plans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutPlan"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workouts"
        ],
        "summary": "Create workout plan",
        "description": "Create a new workout plan",
        "requestBody": {
          "description": "Workout plan data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutPlan"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workout plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutPlan"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/api/workouts/{id}": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Get workout plan by ID",
        "description": "Retrieve a specific workout plan with details",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Unique identifier",
              "example": "123"
            },
            "required": true,
            "description": "Unique identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Workout plan details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutPlan"
                }
              }
            }
          },
          "404": {
            "description": "Workout plan not found"
          }
        }
      },
      "put": {
        "tags": [
          "Workouts"
        ],
        "summary": "Update workout plan",
        "description": "Update an existing workout plan (coach only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Unique identifier",
              "example": "123"
            },
            "required": true,
            "description": "Unique identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "description": "Updated workout plan data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutPlan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workout plan updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutPlan"
                }
              }
            }
          },
          "404": {
            "description": "Workout plan not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Workouts"
        ],
        "summary": "Delete workout plan",
        "description": "Delete a workout plan (coach only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Unique identifier",
              "example": "123"
            },
            "required": true,
            "description": "Unique identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Workout plan deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                }
              }
            }
          },
          "404": {
            "description": "Workout plan not found"
          }
        }
      }
    },
...
  },
...
}